downloader: Move VCS related tests to a sub-directory for consistency
The package names mostly already contained a ".vcs" sub-package.
